add Person

+++ samples/spark-java-samples/src/main/java/com/waylau/spark/java/samples/common/Person.java
@@ -2,44 +2,30 @@ package com.waylau.spark.java.samples.common;
 
 public class Person {
     private String name;
-
-    private Integer age;
-
+    private Long age;
     private String homePage;
 
     public String getName() {
-
         return name;
-
     }
 
     public void setName(String name) {
-
         this.name = name;
-
     }
 
-    public Integer getAge() {
-
+    public Long getAge() {
         return age;
-
     }
 
-    public void setAge(Integer age) {
-
+    public void setAge(Long age) {
         this.age = age;
-
     }
 
     public String getHomePage() {
-
         return homePage;
-
     }
 
     public void setHomePage(String homePage) {
-
         this.homePage = homePage;
-
     }
 }
